api/requirements.txt
fastapi>=0.100.0
uvicorn==0.22.0
motor==3.1.2
pydantic>=2.0.0
email-validator==2.0.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1
